fix decaffeinated 'for' loops
Looks like #113 and #116 are the result of a weirdness in some `for` loops
(probably when automatically converting code from coffee script to javascript ?).

Anyway, it looks like the intention was to allow negative as well as positive
sequences in the `for` condition. However, if this was the case the
final step needs to be changed.

```
- for (var i = 0; 0 < val ? i <= val : i >= val; 0 < val ? i++ : i++)
+ for (var i = 0; 0 < val ? i <= val : i >= val; 0 < val ? i++ : i--)
```

This commit is effectively a (manual) find and replace of the above typo
correction in all source files.