fix decaffeinated 'for' loops
Looks like #113 and #116 are the result of a weirdness in some `for` loops
(probably when automatically converting code from coffee script to javascript ?).

Anyway, it looks like the intention was to allow negative as well as positive
sequences in the `for` condition. However, if this was the case the
final step needs to be changed.

```
- for (var i = 0; 0 < val ? i <= val : i >= val; 0 < val ? i++ : i++)
+ for (var i = 0; 0 < val ? i <= val : i >= val; 0 < val ? i++ : i--)
```

This commit is effectively a (manual) find and replace of the above typo
correction in all source files.


Former-commit-id: ed03a38a750de78fd7e32d31bd402e8e79e0feaf [formerly aab0b9fc380596c36a5c0a48c5335ebd7ea24a71]
Former-commit-id: 505f7c72fe9f53687f7d43e6729b3a05eaae61e0