add options for colouring conservation scores
Useful to have some more options to specify how the
conservation scores are displayed.

Adds the following options to the main config:

```
conserv: {
	maxHeight: 20,
	fillColor: ['#660', '#ff0'],
	strokeColor: '#000'
}
```

This lets the user have more control of the appearance of the
conservation scores. The height is a number in pixels. The colours
can be specified as anything that D3 considers a colour e.g. "#00F",
"rgb(0,0,1)", "blue" (see https://github.com/d3/d3-color)

For `fillColor` and `strokeColor`:

 - If a single value is given then all the elements are coloured the same

 - If a [min,max] range is given then each element is coloured according to
the height

See: ```snippets/conserv```

NOTE: this adds d3-scale as a dependency which seems to add ~25K to the
min.gz. I would argue that this is worth it as being able to interpolate
(and translate) colors is pretty useful generally.